Module plotters_bitmap::bitmap_pixel

Structs

BGRXPixel

The marker type that indicates we are currently using a BGRX8888 pixel format

RGBPixel

The marker type that indicates we are currently using a RGB888 pixel format

Traits

PixelFormat

The trait that describes some details about a particular pixel format